+++ Infrastructure/WhitePage.ResourceAccess/Contracts/Core/ICommonDataAccess.cs
@@ -14,9 +14,9 @@ namespace WhitePage.ResourceAccess.Contracts.Core
         List<KeyValuePair<string, KeyValuePair<string, int>>> GetCenterWiseChartObjectValues(int id);
         List<KeyValuePair<string, KeyValuePair<string, int>>> GetCounselorWiseChartObjectValues(int id);
         List<KeyValuePair<string, KeyValuePair<string, int>>> GetPeacemakerWiseChartObjectValues(int id);
-        List<KeyValuePair<string, KeyValuePair<string, int>>> GetCenterWiseChartObjectValues(string column);
-        List<KeyValuePair<string, KeyValuePair<string, int>>> GetCounselorWiseChartObjectValues(string column);
-        List<KeyValuePair<string, KeyValuePair<string, int>>> GetPeacemakerWiseChartObjectValues(string column);
+        List<KeyValuePair<string, KeyValuePair<string, double>>> GetCenterWiseChartObjectValues(string column);
+        List<KeyValuePair<string, KeyValuePair<string, double>>> GetCounselorWiseChartObjectValues(string column);
+        List<KeyValuePair<string, KeyValuePair<string, double>>> GetPeacemakerWiseChartObjectValues(string column);
         List<KeyValuePair<string, KeyValuePair<string, float>>> GetCenterWiseAvgChartObjectValues(string column);
         List<KeyValuePair<string, KeyValuePair<string, float>>> GetCounselorWiseAvgChartObjectValues(string column);
         List<KeyValuePair<string, KeyValuePair<string, float>>> GetPeacemakerWiseAvgChartObjectValues(string column);
